+++ CHANGELOG.md
@@ -53,3 +53,4 @@
 # 0.1.6
 * Define `->>>`, `->>>>`, `->>>>>` methods for `Traversable` to compact expressions
 * Move `Natural` transformation from `Functor` to `Transformation` module
+* Define infix and infix flipped versions of methods of `Functor` typeclasses